Only render what's seen; change structure
Not like the rendering speed matters at this point, but we shouldn't
need to worry about rendering performance at all now unless much more
complex effects are added in. This laptop gets 2000+ FPS now regardless
of map size, where it got 631 before on a 128x128 map and ~180 on a
256x256 map. The individual pieces are stored differently, and include a
name and a fixed palette instead of depending on guiRandom giving the
same number in each render call. Further performance improvements are
probably possible, but I don't think they're needed yet.