Moving pieces! Removed Android project for now
The pieces now move all around the map, disregarding terrain. Writing
this discovered multiple bugs across SquidLib and here, so it was not
especially easy. The Android project is out of here for now, because
there are issues with the build and Gradle not playing nicely (even for
running the game as a desktop application, it tried to use an invalid
Android installation... it's better with it gone until the plugins
mature).