Reduce size of texture, text is working, world gen
The world is generated a little differently now, and it seems to be a
bit better. Text displays correctly now; before, the texture had the
wrong color for text so it all looked dark blue. The image texture (that
also includes the font) is also, somehow, much smaller as a PNG8 than
pngout could shrink it by as, I believe, a PNG24.